Move PlayerInventory to own file
