+++ src/main/kotlin/game/player/InventoryItem.kt
@@ -1,0 +1,7 @@
+package game.player
+
+enum class InventoryItem {
+    ARROW,
+    FOOD,
+    GOLD,
+}
