+++ src/main/kotlin/game/player/PlayerInventory.kt
@@ -1,0 +1,7 @@
+package game.player
+
+data class PlayerInventory(private val inventoryItems: Map<InventoryItem, Int> = mapOf()) {
+    fun getInventory() = inventoryItems
+    fun copyThis(inventoryItem: Map<InventoryItem, Int> =
+                         this.getInventory()) = PlayerInventory(inventoryItem)
+}
